requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
dig==0.1.10
Django==3.0.3
django-appconf==1.0.4
django-avatar==5.0.0
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-imagekit==4.0.2
docopt==0.6.2
et-xmlfile==1.0.1
Faker==4.1.0
fastcache==1.1.0
futures==3.1.1
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
jdcal==1.4.1
mpmath==1.1.0
num2words==0.5.10
oauth2client==4.1.3
openpyxl==3.0.3
pilkit==2.0
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.7.0
rsa==4.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
sympy==1.5.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9
uuid==1.30
wincertstore==0.2
yagmail==0.11.224
